Pulling changes from moonlight upstream
https://github.com/moonlight-stream/moonlight-chrome.git
up to
643cc6d128c74927e665307eb3cd9ad070d868b8

Also updating
https://github.com/moonlight-stream/moonlight-common-c
up to commit:
c8faf3539b0c33b9e5eb8f5cf99c04d307c0e608

Bug: https://cam.sprc.samsung.pl/browse/VDWASM-169

+++ wasm/moonlight_wasm.hpp
@@ -110,9 +110,9 @@ class MoonlightInstance {
   static void* StopThreadFunc(void* context);
 
   static void ClStageStarting(int stage);
-  static void ClStageFailed(int stage, long errorCode);
+  static void ClStageFailed(int stage, int errorCode);
   static void ClConnectionStarted(void);
-  static void ClConnectionTerminated(long errorCode);
+  static void ClConnectionTerminated(int errorCode);
   static void ClDisplayMessage(const char* message);
   static void ClDisplayTransientMessage(const char* message);
   static void ClLogMessage(const char* format, ...);
@@ -292,4 +292,4 @@ EM_BOOL handlePointerLockError(int eventType,
                                void *userData);
 
 void onConnectionStarted();
-void onConnectionStopped(long errorCode);
+void onConnectionStopped(int errorCode);
